[draft][clang codegen] Fix C23 "empty initialization".
C23 rules for "empty initialization" require that, unlike initializers
that are explicitly written out, all padding must be initialized to
zero.  So we need to use different code sequences.

There are two big gaps which are not easy to solve due to the
representation:

- APValue has no way to represent zeroed padding
- An LLVM IR "zeroinitializer" has no way to represent zeroed padding.

For LLVM IR, we can rewrite the types we use for such initializers;
there's no fundamental difficulty, but it's slightly tricky code, and
we probably need a ton of test updates if we touch default
initialization.

Not sure what the solution looks like for APValue.