[WIP][AArch64] Support CodeView debug info with SVE.
From quick testing, there are two major things that need to be added to
make the debug info work: one, we need to encode the register numbers
themselves into the debug info, and two, we need to be able to encode
the offsets of SVE variables on the stack.

This patch adds hacks for both, as a demonstration of what's required.
Hopefully we can do better... but this is better than nothing.